git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2125 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/sgi.c
@@ -1097,7 +1097,7 @@ static MagickBooleanType WriteSGIImage(const ImageInfo *image_info,Image *image)
           Write out line start and length tables and runlength-encoded pixels.
         */
         for (i=0; i < (ssize_t) (iris_info.rows*iris_info.depth); i++)
-          (void) WriteBlobMSBLong(image,(size_t) offsets[i]);
+          (void) WriteBlobMSBLong(image,(unsigned int) offsets[i]);
         for (i=0; i < (ssize_t) (iris_info.rows*iris_info.depth); i++)
           (void) WriteBlobMSBLong(image,(unsigned int) runlength[i]);
         (void) WriteBlob(image,number_packets,packets);
